New Parser module
Moved parser functions in the Rule class to it's own module. This means that most of the moving parts in the Rule have been encapsulated in different modules. I have to be careful about how this impacts the flow of dependencies but so far everything runs fine.
